Best practices
Added static and template folders for frontend.
Added api_routes.py for backend.

Both are best practices.
