--- src/backend/config.py
@@ -1,1 +1,0 @@
-OPENAI_API_KEY=""
