+++ src/frontend/static/script.js
