+++ src/frontend/templates/index.html
@@ -13,6 +13,6 @@
     <h2>Generated Case:</h2>
     <p id="case-output"></p>
 
-    <script src="script.js"></script>
+    <script src="static/script.js"></script>
 </body>
 </html>
